* MS search panel accepts node-ids and node-revisions, in which case it just displays the matching node. * MS you can link an orphaned node into a new location. * Fixed that single-premise arg with no premise, but with relevance arguments, would not show the "(single-premise arg lacks base-claim; right-click to add)" message. (thus blocking ability to add the single premise!)

+++ Source/Server/Commands/LinkNode_HighLevel.ts
@@ -63,7 +63,7 @@ export class LinkNode_HighLevel extends Command<Payload, {argumentWrapperID?: st
 		this.node_data = GetNodeL2(nodeID);
 		AssertV(this.node_data, "node_data is null.");
 		const oldParent_data = GetNodeL2(oldParentID);
-		AssertV(oldParent_data, "oldParent_data is null.");
+		//AssertV(oldParent_data, "oldParent_data is null."); // commented: allow linking orphaned nodes
 		this.newParent_data = GetNodeL2(newParentID);
 		AssertV(this.newParent_data, "newParent_data is null.");
 
@@ -113,7 +113,7 @@ export class LinkNode_HighLevel extends Command<Payload, {argumentWrapperID?: st
 			this.sub_unlinkFromOldParent.Validate();
 
 			// if the old parent was an argument, and the moved node was its only child, also delete the old parent
-			if (deleteOrphanedArgumentWrapper && oldParent_data.type === MapNodeType.Argument && oldParent_data.children.VKeys().length === 1) {
+			if (deleteOrphanedArgumentWrapper && oldParent_data && oldParent_data.type === MapNodeType.Argument && oldParent_data.children.VKeys().length === 1) {
 				this.sub_deleteOldParent = this.sub_deleteOldParent ?? new DeleteNode({mapID, nodeID: oldParentID}).MarkAsSubcommand(this);
 				this.sub_deleteOldParent.childrenToIgnore = [nodeID]; // let DeleteNode sub that it doesn't need to wait for nodeID to be deleted (since we're moving it out from old-parent simultaneously with old-parent's deletion)
 				this.sub_deleteOldParent.Validate();
